requirements.txt
typer==0.12.5
rich==13.9.4
click==8.1.7
pymongo>=4.6.0
python-dotenv>=1.0.0
